+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - iOS/Mac: More detailed errors on invalid status codes. #263 #232
 - iOS: Prevent NETunnelProviderSession initialization after disconnect
+- iOS/Mac: Update dependencies to latest versions. #273
 
 ## 2.2 (2020-05-01)