+++ CHANGES.md
@@ -4,6 +4,8 @@
 
 - iOS/Mac: More detailed errors on invalid status codes. #263 #232
 - macOS: Fix TunnelKit to get IPv6 working in macOS Mojave
+- iOS: Prevent NETunnelProviderSession initialization after disconnect
+- iOS/Mac: Update dependencies to latest versions. #273
 
 ## 2.2 (2020-05-01)