+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- iOS/Mac: Preserve tunnel log across the previous tunnel invocation
 
 ## 2.2 (2020-05-08)